Fix transient memory leak from unwind_protect and safe
R_UnwindProtect always adds the result of the protected function to the
CAR of the continuation token.
See https://github.com/wch/r-source/blob/17b4af6c84200d86c26f8f44e176efe5674a305c/src/main/context.c#L945.

Because we only allocate a single continuation token and put it in the
precious list this implicitly protects the result until the next call to
`unwind_protect`.

However if there is no error and R_UwindProtect does a normal exit the
this memory should not be protected, so we now unset the CAR of the
token before returning the value ourselves. I _think_ it could be argued that the
implementation of R_UwindProtect should do this automatically, but I
don't think this use case of having a long lived token was considered.

This leak was transient, once another cpp11 function calls
`unwind_protect` the old value in the CAR will be overwritten and the
memory would be freed, but it is still nice to have it fixed.

Fixes #154