Return a zero length vector default constructed are converted to SEXP
When we default construct writable vectors we set them to R_Nilvalue, so
the default constructor cannot throw an exception. However this causes
problems if you then immediately return them. By checking for this
condition in the SEXP conversion function we can return a 0 length
vector in these cases, which is what the user likely intended.

Fixes #166