Reverts #278
CRAN advised that we revert this since the "Classic" intel compiler this was intended to support is deprecated and will be removed in 2023.

It also did not even really help the original issue (https://github.com/r-lib/systemfonts/issues/98) since there were other compilation errors there.

Since this is such an old system and we have only ever had 1 report about this, I feel fine reverting it back.