Use `vector_(rhs.vector())` rather than `vector_(rhs)`
The latter goes from `matrix -> SEXP -> r_vector` in a way that always avoids a copy, regardless of read-only or writable. This new way ensures that we go `r_vector -> r_vector`, and the correct `r_vector` constructor can get called. If its a writable input, it will be correctly copied now.