Use `Rf_shallow_duplicate()` in `r_vector(const r_vector& rhs)` constructor after all
The fact that `reserve_data()` drops `dim` and `dimnames` unconditionally is problematic here, because this copy constructor gets called from inside the `matrix` class.

Really we should restrict usage of `reserve_data()` to JUST `reserve()`. This should be fine because it will only get called if we `push_back()`, and you can't `push_back()` on a `matrix`.